refactor: sorrydb/sorrydb to  sorrydb/database
